requirements.txt
aio-pika==6.7.1
aiohttp==3.6.0
aiohttp-cors==0.7.0
aiormq==3.2.3
amqp==5.0.1
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
gunicorn==19.7.1
kombu
motor==2.2.0
multidict==4.7.6
pamqp==2.3.0
PyJWT==1.7.1
pymongo==3.11.0
PyYAML==5.1.2
uvloop==0.14.0
requests==2.25.1
dnspython